requirements.txt
art==5.6
pyEarnapp==0.0.16.1
python-dotenv==0.20.0